Specify the name of the output index file
By default, bcftools index creates the index file in the same place
as the vcf (I think). It looks like we don't have permission to write
to the mounted directory, so we'll have to create it in the current
working directory and then later specify that it is the index for
a given vcf. (We'll get to that later - for now just trying to create
the vcf.)